requirements.txt
# Core dependencies
requests==2.31.0
psycopg2-binary==2.9.9
pandas==2.1.4
python-dotenv==1.0.0

# Data validation
pydantic==2.5.3

# Scheduling
schedule==1.2.1

# Logging
python-json-logger==2.0.7

# Testing
pytest==7.4.3
